feat: support UDAFs with and without init Args with same param type (#5982)
As well as an expression, UDAFs support constants being passed. However, prior to this change it was not possible to register a UDAF that have variants that too no constants and constants.  For example:

```java
@UdafDescription(
    name = "LATEST_BY_OFFSET",
    description = LatestByOffset.DESCRIPTION,
    author = KsqlConstants.CONFLUENT_AUTHOR
)
public final class LatestByOffset {

  ...

  // Version taking no constants:
  @UdafFactory(description = "return the latest value of an integer column",
      aggregateSchema = "STRUCT<SEQ BIGINT, VAL INT>")
  public static Udaf<Integer, Struct, Integer> latestInteger() {
    return latest(STRUCT_INTEGER);
  }

  // Version taking one constant:
  @UdafFactory(description = "return the last N values of an integer column",
      aggregateSchema = "ARRAY<STRUCT<SEQ BIGINT, VAL INT>>")
  public static Udaf<Integer, List<Struct>, List<Integer>> latestIntegers(final int topNSize) {
    return latestN(STRUCT_INTEGER, topNSize);
  }
```

The above would previously have thrown an error on start up, complaining that "A function with the same name and parameter types is already registered" when `latestIntegers` was being registered, as it incorrectly only took the runtime `INTEGER` parameter, ignoring the `INTEGER` constant parameter.

With this change all parameters are taken into account.

Co-authored-by: Andy Coates <big-andy-coates@users.noreply.github.com>